Added watchlist

+++ films/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from api.models import Film
 from django.contrib.auth.models import User
 from user.models import Profile
-from .models import FilmRating
+from .models import FilmRating, FilmWatchlist
 from django.db.models import Avg
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
@@ -11,24 +11,37 @@ def list(request):
     if request.user.is_authenticated():
         add_film_id = request.GET.get('add_film_id', None)
         del_film_id = request.GET.get('del_film_id', None)
+        add_watchlist = request.GET.get('add_watchlist', None)
+        del_watchlist= request.GET.get('del_watchlist', None)
         rate = request.GET.get('rate', None)
         activ_user = get_object_or_404(User, username=request.user)
         activ_profile = get_object_or_404(Profile, user=activ_user)
         
-        if request.method == 'GET' and add_film_id is not None:
+        if request.method == 'GET' and add_film_id is not None and rate is not None:
             film = get_object_or_404(Film, id=add_film_id)
             activ_profile.film.add(film)
-
             if FilmRating.objects.filter(user=activ_user, film=film).exists():
                 filmuser = FilmRating.objects.filter(user=activ_user, film=film)
                 filmuser.update(rate=rate)
             else:
                 FilmRating.objects.create(user=activ_user, film=film, rate=rate)
 
-        if request.method == 'GET' and del_film_id is not None:
+        if request.method == 'GET' and del_film_id is not None and rate is not None:
             film = get_object_or_404(Film, id=del_film_id)
             activ_profile.film.remove(film)
             FilmRating.objects.filter(user=activ_user, film=film, rate=rate).delete()
+        
+        if request.method == 'GET' and add_watchlist is not None:
+            film = get_object_or_404(Film, id=add_watchlist)
+            if FilmWatchlist.objects.filter(user=activ_user, film=film).exists():
+                watchlist_film = FilmWatchlist.objects.filter(user=activ_user, film=film)
+                watchlist_film.update()
+            else:
+                FilmWatchlist.objects.create(user=activ_user, film=film)
+        
+        if request.method == 'GET' and del_watchlist is not None:
+            film = get_object_or_404(Film, id=del_watchlist)
+            FilmWatchlist.objects.filter(user=activ_user, film=film).delete()
 
         p_films = activ_profile.film.all()
         film_list = Film.objects.all().annotate(average_score=Avg('filmrating__rate'))
@@ -36,6 +49,9 @@ def list(request):
         filmrating = FilmRating.objects.all().filter(user=activ_user)
         f_title = filmrating.values_list('film__title', flat=True)
         f_id = filmrating.values_list('film__id', flat=True)
+
+        watchlist = FilmWatchlist.objects.all().filter(user=activ_user)
+        watchlist_title = watchlist.values_list('film__title', flat=True)
         
         page = request.GET.get('page')
         paginator = Paginator(film_list, per_page=10)
@@ -53,6 +69,8 @@ def list(request):
             'p_films': p_films,
             'filmrating': filmrating,
             'activ_profile': activ_profile,
+            'watchlist_title': watchlist_title,
+            'watchlist': watchlist,
         }
         return render(request, 'list/film_list.html', context)
     else:
